+++ dvc/dependency/base.py
@@ -1,5 +1,3 @@
-import re
-
 from dvc.exceptions import DvcException
 
 
@@ -16,71 +14,7 @@ class DependencyIsNotFileOrDirError(DvcException):
 
 
 class DependencyBase(object):
-    REGEX = None
+    IS_DEPENDENCY = True
 
-    PARAM_PATH = 'path'
-
-    def __init__(self, stage, path, info=None):
-        self.stage = stage
-        self.project = stage.project
-        self.url = path
-        self.info = info
-
-    def __repr__(self):
-        return "{class_name}: '{url}'".format(
-            class_name=type(self).__name__,
-            url=(self.url or 'No url')
-        )
-
-    def __str__(self):
-        return self.url
-
-    @classmethod
-    def match(cls, url):
-        return re.match(cls.REGEX, url)
-
-    def group(self, name):
-        match = self.match(self.url)
-        if not match:
-            return None
-        return match.group(name)
-
-    @classmethod
-    def supported(cls, url):
-        return cls.match(url) is not None
-
-    @property
-    def scheme(self):
-        return self.path_info['scheme']
-
-    @property
-    def path(self):
-        return self.path_info['path']
-
-    @property
-    def sep(self):
-        return '/'
-
-    @property
-    def exists(self):
-        return self.remote.exists(self.path_info)
-
-    def changed(self):
-        raise NotImplementedError
-
-    def status(self):
-        if self.changed():
-            # FIXME better msgs
-            return {str(self): 'changed'}
-        return {}
-
-    def save(self):
-        raise NotImplementedError
-
-    def dumpd(self):
-        ret = self.info.copy()
-        ret[self.PARAM_PATH] = self.url
-        return ret
-
-    def download(self, to_info):
-        self.remote.download([self.path_info], [to_info])
+    DoesNotExistError = DependencyDoesNotExistError
+    IsNotFileOrDirError = DependencyIsNotFileOrDirError
