+++ dvc/dependency/gs.py
@@ -1,24 +1,6 @@
-import posixpath
-
-try:
-    from urlparse import urlparse
-except ImportError:
-    from urllib.parse import urlparse
-
-from dvc.dependency.s3 import DependencyS3
-from dvc.remote.gs import RemoteGS
+from dvc.output.gs import OutputGS
+from dvc.dependency.base import DependencyBase
 
 
-class DependencyGS(DependencyS3):
-    REGEX = RemoteGS.REGEX
-
-    def __init__(self, stage, path, info=None, remote=None):
-        super(DependencyGS, self).__init__(stage, path, info=info)
-        self.remote = remote if remote else RemoteGS(stage.project, {})
-        bucket = remote.bucket if remote else urlparse(path).netloc
-        path = urlparse(path).path.lstrip('/')
-        if remote:
-            path = posixpath.join(remote.prefix, path)
-        self.path_info = {'scheme': 'gs',
-                          'bucket': bucket,
-                          'path': path}
+class DependencyGS(DependencyBase, OutputGS):
+    pass
