+++ dvc/output/s3.py
@@ -1,10 +1,12 @@
-from dvc.dependency.s3 import DependencyS3
-from dvc.exceptions import DvcException
+import posixpath
+
+from dvc.utils import urlparse
+from dvc.remote.s3 import RemoteS3
+from dvc.output.base import OutputBase
 
 
-class OutputS3(DependencyS3):
-    PARAM_CACHE = 'cache'
-    PARAM_METRIC = 'metric'
+class OutputS3(OutputBase):
+    REMOTE = RemoteS3
 
     def __init__(self,
                  stage,
@@ -13,41 +15,16 @@ class OutputS3(DependencyS3):
                  remote=None,
                  cache=True,
                  metric=False):
-        super(OutputS3, self).__init__(stage, path, info, remote=remote)
-        self.use_cache = cache
-        self.metric = metric
-        if cache and self.project.cache.s3 is None:
-            raise DvcException("no cache location setup for 's3' outputs.")
-
-    def dumpd(self):
-        ret = super(OutputS3, self).dumpd()
-        ret[self.PARAM_CACHE] = self.use_cache
-        ret[self.PARAM_METRIC] = self.metric
-        return ret
-
-    def changed(self):
-        if super(OutputS3, self).changed():
-            return True
-
-        if self.use_cache \
-           and self.project.cache.s3.changed(self.path_info, self.info):
-            return True
-
-        return False
-
-    def checkout(self, force=False):
-        if not self.use_cache:
-            return
-
-        self.project.cache.s3.checkout(self.path_info, self.info)
-
-    def save(self):
-        super(OutputS3, self).save()
-
-        if not self.use_cache:
-            return
-
-        self.info = self.project.cache.s3.save(self.path_info)
-
-    def remove(self, ignore_remove=False):
-        self.remote.remove(self.path_info)
+        super(OutputS3, self).__init__(stage,
+                                       path,
+                                       info=info,
+                                       remote=remote,
+                                       cache=cache,
+                                       metric=metric)
+        bucket = remote.bucket if remote else urlparse(path).netloc
+        path = urlparse(path).path.lstrip('/')
+        if remote:
+            path = posixpath.join(remote.prefix, path)
+        self.path_info = {'scheme': self.scheme,
+                          'bucket': bucket,
+                          'path': path}
