+++ dvc/output/ssh.py
@@ -1,10 +1,12 @@
-from dvc.dependency.ssh import DependencySSH
-from dvc.exceptions import DvcException
+import posixpath
+
+from dvc.utils import urlparse
+from dvc.output.base import OutputBase
+from dvc.remote.ssh import RemoteSSH
 
 
-class OutputSSH(DependencySSH):
-    PARAM_CACHE = 'cache'
-    PARAM_METRIC = 'metric'
+class OutputSSH(OutputBase):
+    REMOTE = RemoteSSH
 
     def __init__(self,
                  stage,
@@ -13,41 +15,24 @@ class OutputSSH(DependencySSH):
                  remote=None,
                  cache=True,
                  metric=False):
-        super(OutputSSH, self).__init__(stage, path, info, remote=remote)
-        self.use_cache = cache
-        self.metric = metric
-        if cache and self.project.cache.ssh is None:
-            raise DvcException("no cache location setup for 'ssh' outputs.")
+        super(OutputSSH, self).__init__(stage,
+                                        path,
+                                        info=info,
+                                        remote=remote,
+                                        cache=cache,
+                                        metric=metric)
+        host = remote.host if remote else self.group('host')
+        port = remote.port if remote else RemoteSSH.DEFAULT_PORT
+        user = remote.user if remote else self.group('user')
 
-    def dumpd(self):
-        ret = super(OutputSSH, self).dumpd()
-        ret[self.PARAM_CACHE] = self.use_cache
-        ret[self.PARAM_METRIC] = self.metric
-        return ret
+        if remote:
+            path = posixpath.join(remote.prefix,
+                                  urlparse(path).path.lstrip('/'))
+        else:
+            path = self.match(self.url).group('path')
 
-    def changed(self):
-        if super(OutputSSH, self).changed():
-            return True
-
-        if self.use_cache \
-           and self.project.cache.ssh.changed(self.path_info, self.info):
-            return True
-
-        return False
-
-    def checkout(self, force=False):
-        if not self.use_cache:
-            return
-
-        self.project.cache.ssh.checkout(self.path_info, self.info)
-
-    def save(self):
-        super(OutputSSH, self).save()
-
-        if not self.use_cache:
-            return
-
-        self.info = self.project.cache.ssh.save(self.path_info)
-
-    def remove(self, ignore_remove=False):
-        self.remote.remove(self.path_info)
+        self.path_info = {'scheme': 'ssh',
+                          'host': host,
+                          'port': port,
+                          'user': user,
+                          'path': path}
