+++ dvc/remote/azure.py
@@ -7,12 +7,8 @@ try:
 except ImportError:
     BlockBlobService = None
 
-try:
-    from urlparse import urlparse
-except ImportError:
-    from urllib.parse import urlparse
-
 import dvc.logger as logger
+from dvc.utils import urlparse
 from dvc.progress import progress
 from dvc.config import Config
 from dvc.remote.base import RemoteBase
