+++ dvc/remote/gs.py
@@ -5,12 +5,8 @@ try:
 except ImportError:
     storage = None
 
-try:
-    from urlparse import urlparse
-except ImportError:
-    from urllib.parse import urlparse
-
 import dvc.logger as logger
+from dvc.utils import urlparse
 from dvc.remote.base import RemoteBase
 from dvc.config import Config
 from dvc.progress import progress
